Split Models, Extensions, and Constants into separate files per type.
Also added documentation for the Color+Hex extension.
